+++ CHANGELOG.md
@@ -1,1 +1,9 @@
+# Changelog
+## 1.0.2 - 2017-07-07
+- Fix for non default value in hitCount column needed for some database engines
+
+## 1.0.1 - 2017-06-02
+- Added hit count and last hit date functionality
+
+## 1.0.0 - 2017-06-01
 - Initial release